Address most implementation review feedback
