+++ apps/website/src/pages/facilitator-feedback/[groupId]/success.tsx
@@ -5,7 +5,7 @@ import Confetti from 'react-confetti';
 import { ProgressDots } from '@bluedot/ui';
 import { PiCheck, PiCreditCard } from 'react-icons/pi';
 import { generateInvoiceUrl } from '../../../lib/generateInvoiceUrl';
-import { isFlaggedFromAirtable } from '../../../lib/facilitatorFollowUps';
+import { airtableToFollowUps, isFlagged } from '../../../lib/facilitatorFollowUps';
 import { trpc } from '../../../utils/trpc';
 
 const formatNames = (names: string[]): string => {
@@ -71,7 +71,7 @@ const FacilitatorFeedbackSuccessPage = () => {
     : null;
 
   const flaggedNames = data.existingPeerFeedback
-    .filter((pf) => isFlaggedFromAirtable(pf.nextSteps))
+    .filter((pf) => isFlagged(airtableToFollowUps(pf.nextSteps)))
     .map((pf) => pf.recipientName)
     .filter(Boolean);
